more int to size_t conversions, fixing one or more amd64 bugs plus a whitespace patch on config.c from vicman
svn:r2482

+++ trunk/src/or/circuitbuild.c
@@ -350,7 +350,7 @@ int circuit_send_next_onion_skin(circuit_t *circ) {
   int r;
   char payload[2+4+DIGEST_LEN+ONIONSKIN_CHALLENGE_LEN];
   char *onionskin;
-  int payload_len;
+  size_t payload_len;
 
   tor_assert(circ && CIRCUIT_IS_ORIGIN(circ));
